backend - v2 register all senarios/requirements.txt
fastapi
email-validator
uvicorn[standard]
sqlalchemy
pyodbc
passlib[bcrypt]
argon2-cffi
python-multipart
python-jose[cryptography]
python-dotenv
pytest
httpx